Make power radius a float

+++ src/Hazard.h
@@ -78,7 +78,7 @@ public:
 	FPoint speed;
 	float base_speed;
 	int lifespan; // ticks down to zero
-	int radius;
+	float radius;
 	int power_index;
 
 	int animationKind;	// direction or other, it is a specific value according to
